Call LimeSurvey API Directly

General function used internally by every other ls_ fellow. Useful when you want something special that is not (yet) implemented in reschola.

Usage

ls_call(method, params = list())

Arguments

method

character, a method supported by LimeSurvey API.

params

list, arguments of the method. Need to be in order stated in documentation. Note that sSessionKey auth credentials are already provided as the first entry of params list.

Value

A tibble, or raw object if server response cannot be reasonably coerced to a tibble.

Details

The list of available LimeSurvey API calls briefly documents the functionality provided. You must strictly adhere to the arguments positions, as they are passed as an array.

See also

Examples

if (FALSE) { # \dontrun{
ls_call("get_survey_properties", params = list(iSurveyID = 123456))
} # }