Login to LimeSurvey API

Obtains XML-RPC/JSON-RPC session key and stores it in dedicated environment for further use by fellow ls_ functions.

ls_login(
 
    api_url = "https://dotazniky.scholaempirica.org/limesurvey/index.php/admin/remotecontrol"
)

Arguments

api_url

character, URL of the API endpoint, default to SCHOLA EMPIRICA's LimeSurvey API.

Value

No "explicit" return value, but assigns the session key to a dedicated environment.

Details

By default, https://dotazniky.scholaempirica.org/ is used as the LimeSurvey server prividing the API. The credentials used for user authentication are obtained through interactive prompts, mainly for security reasons. The function tries to obtain the credentials from the system environment variable table first. If none found, the user is asked for them and is provided with guidance for permanent credential storage.

See also

Examples

if (FALSE) {
ls_login()
}