Schola Empirica Word document with customisable template

This is a function called in the output part of the YAML section of the Rmd file while using the Word template provided at the same place (see example below).

Usage

schola_word2(...)

Arguments

...

Arguments to be passed to [bookdown::word_document2]

Value

A modified word_document2 with the standard Schola formatting, but without hard-coded and unchangeable template.

Details

Compared to schola_word, this "version" comes with no predefined template, so the user can utilize a template stated in YAML header (see the example below, or read the bookdown manual for more details and for a brief guide to Word templating).

See also

Author

Jan Netik

Examples

if (FALSE) { # \dontrun{
output:
reschola::schola_word2:
reference_docx:template.docx
} # }