Plot standard Schola likert-like barplot with groupwise comparison per items

TODO description...

Usage

schola_barplot(
  .data,
  vars,
  group,
  dict = dict_from_data(.data),
  escape_level = "nevím",
  n_breaks = 11,
  desc = TRUE,
  labels = TRUE,
  min_label_width = 0.08,
  absolute_counts = TRUE,
  fill_labels = waiver(),
  ...
)

Arguments

.data

data with items and group variable

vars

vector of items, supports tidyselect syntax (i.e, non-standard evaluation)

group

group variable used to split the results, have to be logical, where TRUE is gonna be considered as "focal" group and displayed as upper group with full opacity. Supports tidyselect syntax.

dict

item code-label dictionary, if none provided, those are derived from the data.

escape_level

character, level of item response considered as NA

n_breaks

number of breaks displayed at x-axis, outer labels are automatically aligned to face inward. Defaults to 11, which results in 10% wide breaks.

desc

sor items in descending order?

labels

draw labels?

min_label_width

minimal label width that is displayed in the plot

absolute_counts

draw labels and absolute counts in parentheses?

fill_labels

character vector or function taking breaks and returning labels for fill aesthetic

...

Arguments passed on to fct_nanify

negate

logical, whether to return non-matching elements. Defaults to FALSE.

ignore_case

logical, ignore case when matching? Defaults to TRUE.

Value

object of class "gg", "ggplot"

Details

TODO details...